feat(#236): 設定を再編（UI ページ新設・試験機能の卒業）
- テーマ・言語を新設の［ユーザーインターフェイス］ページに切り出す（［全般］直後）。
- 試験機能の非推奨以外（外部ブラウザー関連）を［全般］の［外部ブラウザー］
  セクションに集約して卒業させる:
  新規投稿/タイムスタンプを外部ブラウザーで開く、リンクを開くブラウザー、
  Edge プロファイル。Edge プロファイルは「リンクを開くブラウザー」の従属
  オプションとして Expander 内に格納。
- ［試験機能］ページには非推奨項目（定期アクティブ化・状態表示）のみ残す。
- リソース Nav_UserInterface（ja/en）を追加。
