+++ Strings/en-US/Resources.resw
@@ -113,6 +113,7 @@
 
   <!-- Settings window navigation -->
   <data name="Nav_General" xml:space="preserve"><value>General</value></data>
+  <data name="Nav_UserInterface" xml:space="preserve"><value>User Interface</value></data>
   <data name="Nav_Experimental" xml:space="preserve"><value>Experimental</value></data>
   <data name="Nav_Extensions" xml:space="preserve"><value>Extensions</value></data>
   <data name="Nav_Profiles" xml:space="preserve"><value>Profiles</value></data>
